New config option for tmux attach
If you run tmux attach and there is a session, then it will attach to
that session (whether it's already attached or not). If there is not
a session already then it will create one for you.

Was needed because the old way was not working with mosh.

via: https://unix.stackexchange.com/questions/103898/how-to-start-tmux-with-attach-if-a-session-exists/103956#103956